LayoutTests/editing/selection/select-out-of-floated-non-editable-01.html

<!DOCTYPE html>
<html>
<body>
<div id="container"><div>block-level text before floats</div><div id="target" style="float: left; margin-right: 20px">first float*</div><div style="float: left; margin-right: 20px">second float*</div>inline-level text after the floats</div>
<p style="padding-top: 1em;">
To test manually, drag from the middle of the 'first float' div to the
right, into the space between floats. The selection should go to the end
of the 'second float' div with a portion of the inline-level box's
background and not jump to the beginning.
</p>
<pre id="console"></pre>
<script src="../../resources/dump-as-markup.js"></script>
<script src="resources/select-out-of-floated-non-editable.js"></script>
<script>
runTest("float*second float*", 10);
</script>
</body>
</html>
