Safari-609.1.11.1/LayoutTests/editing/selection/select-out-of-floated-non-editable-01-expected.txt

block-level text before floats
first float*second float*inline-level text after the floats
To test manually, drag from the middle of the 'first float' div to the right, into the space between floats. The selection should go to the end of the 'second float' div with a portion of the inline-level box's background and not jump to the beginning.

SUCCESS: Selected test is 'float*second float*' as expected
| <div>
|   "block-level text before floats"
| <div>
|   id="target"
|   style="float: left; margin-right: 20px"
|   "first <#selection-anchor>float*"
| <div>
|   style="float: left; margin-right: 20px"
|   "second float*"
| "<#selection-focus>inline-level text after the floats"
